+++ src/components/Main.js
@@ -1,6 +1,5 @@
 import React from 'react'
-import LessonItem from './LessonItem'
-import LessonModal from './LessonModal'
+import ContactForm from './ContactForm'
 
 const lessons = [
   {
@@ -35,11 +34,6 @@ export default class Main extends React.Component {
     this.state = {
       isModalOpen: false,
       openLesson: null,
-      email: '',
-      inquiry: '',
-      emailError: '',
-      inquiryError: '',
-      formSubmitted: false,
     }
   }
 
@@ -54,111 +48,11 @@ export default class Main extends React.Component {
     this.setState({isModalOpen: false})
   }
 
-  handleEmailFocus(event) {
-    const {email} = this.state
-    if (email === '') {
-      this.setState({emailError: '記入必須項目です'})
-    }
-  }
-
-  handleEmailChange(event) {
-    const email = event.target.value
-    if (email === '') {
-      this.setState({
-        email: email,
-        emailError: '記入必須項目です',
-      })
-    } else {
-      this.setState({
-        email: email,
-        emailError: '',
-      })
-    }
-  }
-
-  handleInquiryChange(event) {
-    const inquiry = event.target.value
-    if (inquiry === '') {
-      this.setState({
-        inquiry: inquiry,
-        inquiryError: '記入必須項目です',
-      })
-    } else {
-      this.setState({
-        inquiry: inquiry,
-        inquiryError: '',
-      })
-    }
-  }
-
-  handleInquiryFocus(event) {
-    const {inquiry} = this.state
-    if (inquiry === '') {
-      this.setState({inquiryError: '記入必須項目です'})
-    }
-  }
-
-  handleSubmit(event) {
-    event.preventDefault();
-    this.setState({formSubmitted: true})
-  }
-
-  isFormValid() {
-    const {
-      email,
-      inquiry,
-      emailError,
-      inquiryError,
-    } = this.state
-    return email !== '' && inquiry !== '' && emailError === '' && inquiryError === ''
-  }
-
   render() {
     const {
-      formSubmitted,
-      emailError,
-      email,
-      inquiryError,
-      inquiry,
       isModalOpen,
       openLesson,
     } = this.state
-
-    const isFormValid = this.isFormValid()
-
-    let formJSX = null
-    if (formSubmitted) {
-      formJSX = (
-        <div>ご回答ありがとうございました</div>
-      )
-    } else {
-      formJSX = (
-        <form onSubmit={event => this.handleSubmit(event)}>
-          <p>メールアドレス（必須）</p>
-          <div className='error-message'>{emailError}</div>
-          <input
-            value={email}
-            onFocus={event => this.handleEmailFocus(event)}
-            onChange={event => this.handleEmailChange(event)}
-          />
-          <p>お問い合わせ内容（必須）</p>
-          <div className='error-message'>{inquiryError}</div>
-          <textarea
-            type='text'
-            value={inquiry}
-            onFocus={event => this.handleInquiryFocus(event)}
-            onChange={event => this.handleInquiryChange(event)}
-          />
-          <p>※必須項目は必ずご入力ください</p>
-          <input
-            className={`contact-submit ${!isFormValid ? 'contact-submit-disabled' : ''}`}
-            type='submit'
-            disabled={!isFormValid}
-            value='送信'
-          />
-        </form>
-      )
-    }
 
     return(
       <div className='main'>
@@ -170,24 +64,34 @@ export default class Main extends React.Component {
           <h3 className='section-title'>学べるレッスン</h3>
           {lessons.map(lesson => {
             return (
-              <LessonItem
+              <div
+                className='contents-item'
                 key={lesson.id}
-                lesson={lesson}
-                handleClickImage={event => this.handleClickImage(event, lesson)}
-              />
+              >
+                <img
+                  src={lesson.imageUrl}
+                  onClick={event => this.handleClickImage(event, lesson)}
+                />
+                <p>{lesson.name}</p>
+              </div>
             )
           })}
         </div>
         {isModalOpen && (
-          <LessonModal
-            lesson={openLesson}
-            handleClickClose={event => this.handleClickClose(event)}
-          />
+          <div className='modal-wrapper'>
+            <div className='modal'>
+              <i
+                className='modal-close fas fa-times'
+                onClick={event => this.handleClickClose(event)}
+              />
+              <div className='modal-description'>
+                <h2>{openLesson.name}</h2>
+                <div>{openLesson.description}</div>
+              </div>
+            </div>
+          </div>
         )}
-        <div className='contact-form'>
-          <h3 className='section-title'>お問い合わせ</h3>
-          {formJSX}
-        </div>
+        <ContactForm />
       </div>
     )
   }
